Prevent early free of an ESource when it has pending operations
It could happen, in some situations, that an ESource could be freed
in a dedicated thread, while it had been in use in an idle callback.
As it held a corresponding mutex, the crash happened in the finalize()
function, when it had been tried to free that locked mutex.

The change removes obsolete members from the private structure of
the ESource and references itself for the idle callbacks. Even it
avoids early free (and idle callback cancellation in the dispose()),
it will make sure that the ESource will not disappear while the idle
callback is running, the same as it'll avoid some other races around
this situation.

It had been reported downstream at:
https://bugzilla.redhat.com/show_bug.cgi?id=1533442