[IMAPx] Select destination mailbox only when permanentflags not known yet
The previous code meant that whenever a UID COPY/MOVE had been issued
a SELECT to the destination folder and then (eventually back) to the source
folder was done as well, even when it was not needed. Doing the destination
folder SELECT only when needed speeds up message filtering by skipping
unnecessary SELECT commands.