Replace e_cal_recur_nth array with e_cal_recur_get_localized_nth()
The usage of e_cal_recur_nth array could properly work only if either
the user of it used g_dgettext() with "evolution-data-server" domain,
or the domain in use was the "evolution-data-server" (for bare _()),
or the domain in use had translations for the items in the array, which
makes the translation here too complicated. The new function always
translates in the correct translation domain, thus avoids all the above
issues and also allows to check for array boundary.