Add book/calendar-factory argument to override backend-per-process compile-time option
The evolution-addressbook-factory and evolution-calendar-factory processes
have now an argument --backend-per-process, which can be any integer, where
value 0 means to not use backend-per-process feature, value 1 means to use it
and any other value means to use as evolution-data-server had been configured
during compile-time. The default is to respect the compile-time option.