build: favor python3 as interpreter
Various distros try to minimize the use of python2. Since all our
python scripts are already compatible with python3, we now favor
python3 as interpreter, falling back to python2 as needed.

Also, gen-western-table.py is only used during build and we call it
explicitly as parameter to the detected python interpreter; as such
it does not require a shebang and we don't have to confuse anybody
looking into it what version python we care for.