Change how next occurrence of a recurring task is found
The previous naive "set some faraway end time" approach was also
a performance hit for often-happen tasks, because e-cal-recur
computes all occurrences in the given time interval first and only
then lets the caller know with the callback, thus, with a long time
interval most of the found occurrences had been useless. The new
approach is adaptable and avoids this performance issue as well.