[module-secret-monitor] Correct runtime warning from secret_service_search_sync()
The evolution-source-registry console could contain a runtime warning:
WARNING **: secret_service_search_sync: must specify at least one attribute to match
which was due to calling secret_service_search_sync() with an empty
hash table of attributes. Adding an 'eds-origin' attribute to the keys
and search for it helps to make the module do what it was supposed to do.