[CalDAV] Race condition on connection object
It could happen that one thread had been working with the connection
while another thread disconnected it, which could lead to a race
condition on the connection object, especially when the code execution
had been within that object.

The change makes sure to either add a reference on the connection
object while working with it or it adds a connection lock which is
acquired for the whole time when the connection object is used.