ESourceRegistry's thread main_context can leak
The problem was that EWebDAVDiscover (used to find out CalDAV calendars/memo
lists/task lists and CardDAV books) wanted to receive a proxy ESource, for
which it requires either ESourceRegistry or ESourceRegistryServer. As
the EWebDAVDiscover API didn't have any such argument it always created its
own ESourceRegistry instance, which had been dropped shortly afterwards. This
drop also meant that the internal GDBus object scheduled its own events on
the ESourceRegistry's thread main_context (to unsubscribe D-Bus signals), but
this main_context didn't have enough time to flush the queue, which resulted
in a leak of the main_context, because the events in it referenced it (similar
to circular reference). The fix was to flush the queue before closing
the thread and to reorganize the ESourceRegistry's dispose() function, thus
the flush is done on filled main_context.

While I've been in it, some structures unnecessarily long held a GCancellable
object, which also means a file descriptor, thus I fixed that as well.

As the last, but not least, I added a new API for the EWebDAVDiscover to
be able to provide also a function to reference an ESource, thus there is
no need to create the ESourceRegistry instance, thus it is quicker and more
effective now.

Reported downstream at:
https://bugzilla.redhat.com/show_bug.cgi?id=1600489