Make tests retry when factory doesn't know about added ESource yet
When the test creates a new ESource and saves it, then it waits
for an "source-added" signal on an ESourceRegistry instance to know
that it had been "registered" in the registry. It could sometimes
happen that the client received this signal before the factory,
which means that trying to open this ESource fails with a NOT_FOUND
error due to the factory's ESourceRegistry instance wasn't updated yet.
This change tries again for three times with a one second delay, which
should be enough to have the factory's list of known sources updated.