Drop gir and vala bindings generator for libedata-cal
To have it fully working it requires libebackend (otherwise there
are unresolved types, though there are also some from libedataserver),
but gir fails to work due to libebackend contains a dynamic type
(ECacheReaper) and aborts build on it. Maybe once this is fixed,
and all the other quirks, it can be returned back.

There will be the client side bindings only for now.