Make the OAuth2 credentials prompt window default size a bit larger
Just large enough to show whole login prompt of the Google server,
at least as it looks like today. It may not be enough when they improve
it again, but it still should be better than until this change, because
the default size showed only a little portion of the web page, especially
without the "Next" button there, which could be confusing to the users.
The larger default size of the window may make things better also for
the other OAuth2 providers.