M!29 - WebDAV: Permit implicit creation of VAVAILABILITY components
During the creation of new CalDAV calendars, Evolution offers as components
Tasks, Events and Memos.  The user has to select some of the offered components
in order to create the WebDAV collection and the server is asked to create
exactly those components, that the user has chosen.

With this fix, if the user selects all three components, the software assumes
that the user wants a collection supporting whatever components the server
offers, including VTIMEZONE and VAVAILABILITY.

Closes https://gitlab.gnome.org/GNOME/evolution-data-server/merge_requests/29