M!31 - EWebDAVDiscover: focus OK button, when an item is selected
The idea is that once an item from the list is selected, the user
can press Enter and the dialog will be closed.  With this change,
the Address books discovery works as intended, and for Calendars
discovery there is no difference.

The dialog allows in theory selecting more than one entry from the
list, but Evolution does not make use of this option.

Closes: https://gitlab.gnome.org/GNOME/evolution-data-server/merge_requests/31