EDataFactory: Free backend when no client connects to it
This did not work properly with backend-per-process disabled, thus backends
could be left running even when no client had them opened, which could
prevent auto-close of the factory process.

The fix contains also a change no avoid circular reference between the backend
and its view, where view added its own reference to the associated backend and
the backend added a reference to the view when it had been added into
the backend. This could prevent free of the backend when any view was not
properly freed by the client.