GOA module: Run EWS autodiscovery asynchronously
The autodiscovery does network I/O, which means it can block. The place
where it is invoked is called from the main thread of the source registry
process, which means, in case the server is unresponsive for whatever
reason, the blocking call blocks also the source registry, making it
unresponsive until the autodiscovery timeouts or others eds. To fix this,
run the autodiscovery asynchronously instead.