I#196 - Use default OAuth2 services on scratch sources
When creating a new calendar or such, the ESource is a scratch source, which
means it doesn't have its counter part on the evolution-source-registry side.
The OAuth2 tokens are usually received through the source registry, which resulted
in an error "Data source “Personal” does not support OAuth 2.0 authentication".
The sources created in Evolution do not need any special OAuth2 provider,
thus it's possible to use the built-in services only.

Closes https://gitlab.gnome.org/GNOME/evolution-data-server/-/issues/196