ESoupAuthBearer: Shorten the token expiry time by 4 more seconds
Consider the token expired 5 seconds before the server reported it to be expired,
instead of only one second before it, to avoid timing race, which can send
a request with no Bearer Authorization header when the token expires just
between starting the request and sending it. Also pass at least the outdated
token instead of no token at all.