Camel: Fails to filter unread junk messages in a virtual Junk folder
There used to be used the 'read' column to pick read/unread messages,
but this column is used differently, it doesn't reflect the 'seen'
flag in the 'flags' for junk and deleted messages, because it's used
for the unread counts on the folders. Use the 'flags' column directly,
instead of the 'read' column, to get proper value.