ESourceWebDAV: Fallback to SHA1 on SSL trust verification if needed
When a Flatpak application compiles against the evolution-data-server,
which still uses SHA1, but the host version uses SHA256, then the verification
always fails due to SHA version mismatch. It's because the hash to store
is computed by the client (inside the Flatpak), but the connection itself,
with the hash verification, is done on the host.

This covers only the case with older version of the data server in Flatpak.
The opposite case would require a change on the host system.