I#335 - CamelFolder: Calls Junk/Not-Junk learn repeatedly
The JUNK_LEARN flag can change while the job is waiting to be processed,
thus it should recheck whether the learn is still needed and if not,
then skip that message.

The message should be also skipped when it is not in the folder, which
can cause a "No such message" error, for example when marking message
as not junk in On This Computer/Junk folder in some situations.

Also change the arrays of the UIDs, to use the Camel's string pool and
to auto-free the members, which simplifies the new code and saves
memory allocations.

Closes https://gitlab.gnome.org/GNOME/evolution-data-server/-/issues/335