M!83 - Adjust .pc file not to expose used crypto library
Public headers of Evolution Data Server (and Camel) are not directly
including headers of the crypto library NSS. The shared libraries
already have links so NSS is at best only needed for static linking,
but that must go into a the field called "private". The same applies
for SQLite in Camel.

Closes https://gitlab.gnome.org/GNOME/evolution-data-server/-/merge_requests/83