IMAPx: Avoid creating NOSELECT CamelFolder-s
When a folder is set as NOSELECT, it should not be instantiated
as a CamelFolder object, because the folder may not even exist
on the server. The instance also creates a table in the folders.db
database with no content.

The #373 has a `Folders` folder, whose name clashes with Camel's
`folders' table consisting of the information about the stored folders.

Related to https://gitlab.gnome.org/GNOME/evolution-data-server/-/issues/373