Add ECacheKeys helper object for ref-counted data stored in an ECache table
The ECacheKeys is used to store key-value data with a reference count
in a dedicated ECache table. Its code had been already used to store
timezones from the calendar components, this is an extraction of that
code to a new object, which can be re-used for other data.

Related to https://gitlab.gnome.org/GNOME/evolution/-/issues/1245