OAuth2: Show errors returned by the server in the prompt window
The OAuth2 wizard can end with an empty web view, not showing any clue
what is wrong, only in the current URL, where it's too easy to overlook
the real reason why the OAuth2 wizard doesn't continue.

This change shows the error message in the web view/GUI. It also adds
default implementations for the authorization code and an error message
extraction from the URI, to avoid code duplication.