IMAPx: Folder refresh could remove new message from local summary
It could happen that while the folder had been refreshing, another part
added a new message to it, like a filter copying the message, which
the refresh part evaluated as the new message in the summary had been
removed, because it was not part of the 'known_uids', thus it had
been removed from the local summary. The message was still available
in the folder on the server, thus the next time the folder was refreshed
it re-added the new messages back to the local summary.