EReminderWatcher: Last-notified time not always set
There was a bug in the code, expecting the ESource always having
set a valid last-notified time, without which the time was not updated,
but that does not work when it is a fresh new source, with no reminder
shown yet. It's especially visible for read-only calendars, where
the reminder discard fails on "permission denied" error.