EOAuth2ServiceOutlook: Cannot be used with IMAP
The "Outlook" OAuth2 service was not recognized as usable for IMAP
accounts with "outlook.com" or "hotmail.com" user mails, thus make
it explicit this OAuth2 service can be used for IMAP and SMTP with
outlook.com or office365.com servers. That way the autodiscovery
will set OAuth2 for both IMAP and SMTP automatically.

Related to https://gitlab.gnome.org/GNOME/evolution-ews/-/issues/265