evo-I#2720 - Add e_util_host_is_in_domain helper
Several places across the project check whether a given hostname is a
(sub)domain of a specific domain name, usually to check whether the
service is run by a given provider.  Some of the checks are not strict
enough.

Provide a function to do the check correctly, and switch existing checks
to use this function.

Related to https://gitlab.gnome.org/GNOME/evolution/-/issues/2720