M!159 - Add helper object to talk to intune (Microsoft OAuth2 broker service)
This patch adds infrastructure to communicate with a locally running
microsoft-identity-broker (Intune) via DBus to get a PRT SSO cookies.
This infrastructure is needed to implement support for the ms-oapxbc
oauth2.0 protocol extension. Support for that is needed to access
EWS endpoints behind "conditional access policies". At a later point
in time the added infrastructure might be moved to a standalone
library. Until this decision is made and we have broad distro
support, the relevant parts are duplicated here.

The code introduces a new dependency to libuuid.

Co-Developed-by: Milan Crha <mcrha@redhat.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>

Related to https://gitlab.gnome.org/GNOME/evolution-data-server/-/merge_requests/159