Camel: Disable photo-loading when used gpg version does not understand it
The sequoia-chameleon-gnupg does not understand the "show-photos" argument claiming:

   gpg: error: Error parsing --verify-options
   gpg: because: Unknown argument "show-photos"

This change disables the photos loading on the fly until the next start
of the application and re-runs the command. It's okay, because the gpg binary
setting is also read only once per the app run.

Users can disable photo loading by running a command:

    gsettings set org.gnome.evolution-data-server camel-cipher-load-photos false

Related to https://gitlab.gnome.org/GNOME/evolution-data-server/-/issues/581