ESoupSession: Check validity of passed-in URI
The libsoup has a macro which checks whether the URI used to create
a SoupMessage is valid and claims a runtime warning on the console
when it is not valid. Unfortunately, the macro is not part of the public
API, thus it cannot be reused here. It currently checks for a filled
host name only, and non-NULL-ness, thus do the same and return a proper
GError when the passed-in URI is not valid.

Discussed at https://gitlab.gnome.org/GNOME/evolution-data-server/-/issues/584