ESourceRegistry: Reject to create an instance when D-Bus service is not available
The GDBus proxy tries to auto-start the services it is created for, but it
does not error out on its own when the auto-start fails. Check whether there
was found any D-Bus service owner and reject to create the instance when not.

It helps in the Evolution, where the Mail part vanishes all the user settings
related to the folders and account layout as soon as it notices the corresponding
accounts are removed. As the state of no evolution-source-registry process
means no ESource-s are reported, it effectively removes every folder/account
setting only because it could not reach the D-Bus service for some reason.