CamelStore: Speed up open folders with virtual Trash or Junk folder
Opening a folder also means adding it to the virtual Trash or Junk folder,
thus it shows deleted/junk messages in it. Adding the folder also rebuilds
the vTrash folder, which is fine when only one folder is opened, but for
search folders set to include "all local folders" it means to open each
folder one after another, causing as many rebuilds of the vTrash folders
as many folders are available for the local account. Postponing the rebuild
speeds up search folder setup significantly and saves plenty of resources.

Related to https://gitlab.gnome.org/GNOME/evolution/-/issues/3164